Mod: modularizes task list.
